consolidated examples and added info on emphases
